TB_DIST_MAPPING、TB_STOWAGE_MAPPINGに配送便集計項目追加 しきい値を小数点ありに変更

+++ sql/TB_BLOCK.sql
@@ -13,7 +13,7 @@ create table TB_BLOCK (
 																		5:天吊　6：棚			*/
 
 	NU_TDUNIT_CNT			integer					not null,	/* 表示器数						*/
-	NU_THRESHOLD			integer					not null,	/* 閾値							*/
+	NU_THRESHOLD			decimal(5, 1)			not null,	/* 閾値							*/
 
 	DT_TEKIYOKAISHI			nchar(8)				not null,	/* 適用開始日					*/
 	DT_TEKIYOMUKO			nchar(8)				not null,	/* 適用無効日					*/
